requirements.txt
streamlit
torch
torchvision
pillow
numpy
faiss-cpu
git+https://github.com/openai/CLIP.git